Remove QtWidgets
Initially I tried extending QTextEdit to implement a rich-text area whose formatting could be modified, which necessitated the inclusion of QtWidgets as QTextEdit is a Widgets component.

Now that I am simply wrapping a QtQuick TextArea in a wrapper that manipulates its document property, including QtWidgets is no longer necessary.