Remove document_handler
document_handler.cpp and document_handler.h are hereby removed and integrated completely in a derived formattable_text_area implementation.

The first act of this implementation is the addition of the format::getMergedCharFormat() function, which is used to get an entire selection's format with some special rules (see its documentation).