Use TextEdit instead of TextArea
TextArea appears to do "too much" of the work for you. One issue I was
having with it was that text could not be selected if the mouse drag
operation was "too slow", i.e. the user did not drag the mouse quickly
enough across the text. This is not user error, however; this is very
likely to happen if the user wishes to select only one character, for
instance, especially if it's a narrow character where precision is
required.

Also, a new attribute has been added to the main.cpp that directs
Windows platforms to use ANGLE. This is necessary to ensure smooth
window resizing behaviour.