Add AboutQt dialog and child components
Sky.Button is a normal Button that changes the cursor to the pointing
cursor and inherits the right Skywriter colors.

Sky.Dialog is basically a reimplementation of Dialog from the
QtQuick.Dialogs namespace that uses Sky.Button.

Sky.Text has some properties like text wrapping conveniently set so it
doesn't have to be set in every newly instantiated Text.

Sky.LinkText supports hyperlinks with the correct formatting and mouse
event handling.

The other QML files are dialogs that have not yet been implemented.