Add settings components, rework QML file structure
This commit includes several essential changes:

- resources/qml and resources/controls consolidated into
resources/qml/[views/controls]
- several utility types extracted into resources/qml/types
- inline settings replaced with dedicated singleton components

The decision to replace inline settings with dedicated singleton
components was made to ensure consistency across all Settings instances.
Normally, the Settings type does not synchronize its properties across
different instances of Settings types with the same category---which is
probably a good thing, as that would be extremely difficult to implement
well.

However, it is inevitable that more than one Settings instance be used
per category across the entire application. If all Settings instances
were placed on the parent ApplicationWindow, this would result in hacky
and arduous lookup operations where a child component would somehow have
to find said parent window. The usage would also not be pretty.

Instead, it was decided to turn the Settings instances into singleton
types instead. This fulfills all requirements and, due to the module
syntax, allows them to be used in an intuitive and pretty way.

In this application, the `qmlRegisterSingletonType` was used to register
the singleton types. This is not strictly necessary. The alternative
would have been to use a qmldir file. I tinkered with qmldir files for a
bit, however, and was unable to get it working quite right. The
necessary lines of code remain the same regardless, so it doesn't make a
huge difference.