Reverse order of StatsBar labels
Since the Row.layoutDirection is set to RightToLeft to ensure that the
labels are right-aligned, the items also must be in the opposite order
to ensure they render in the correct one.

This still feels like a hacky solution to me. Layout direction is
generally used for languages with an inverted flow like Arabic. I don't
know if using it for layout purposes is technically correct or if this
will have adverse side effects for people who do use a right-to-left
language in their system.

That said, I have been unable to find a better solution. RowLayout is
not an option as the labels either fill the entire RowLayout width
(which would be okay except that the spacings in-between the items are
then inconsistent) or, if they're not, do not render at the right side
of the tool bar.

Row itself has no other way to align its items. Perhaps a GridLayout
could serve, but since that would add an additional level of complexity,
I'll only try to engineer a grid solution if it turns out that
right-to-left is not a good idea.