Handle word, page, and paragraph counts
The algorithm used for the word count is not the same as the one used in
skywriter-javafx. For now it simply uses the QTextCursor's internal word
movement mechanics. Further testing is required to determine if this
algorithm is sufficient or if it handles certain words incorrectly
according to the rules I myself consider "correct". For now, however,
the algorithm appears to be extremely fast.