Remove ScrollView
Removing the ScrollView was necessary to ensure the theme's padding
works as expected. You cannot use a ScrollView that wraps all the text
containers and rectangles, because then you'd be scrolling everything,
including the padding (which is supposed to be a fixed margin at the
window edges). You also cannot use a ScrollView inside the padding
container, because then the ScrollBar would be placed inside that
container as well when actually it should be placed at the screen edge.

The only solution is to do away with the ScrollView and implement a
custom ScrollBar that is placed outside the containers. This ScrollBar
can then bind the text area's y property or can just be attached to a
Flickable. Since there was already a Flickable from the previous
implementation, this is the approach I used.