Change StatsBar into a simple Pane
This change was necessitated by a strange horizontal white line
appearing on the ToolBar. I was able to determine that this line
definitely originates from the ToolBar and is related to its "position"
property. When the "position" property is set to "Header", the line
appears above the ToolBar. When it is set to "Footer", it appears below
the ToolBar.

Changing the root element of StatsBar to a Pane (ToolBar's base type)
resolves the problem.