Add TextIterator to iterate over text
The old method of using the QTextCursor to iterate over text was flawed.
For some reason, it seems as though the QTextCursor considers symbols
such as periods words of their own accord, which led to a lot of false
positives. A reimplementation was necessary.

This will also simplify the implementation of comment exclusion.