Improve ProgressTracker's file stream operations
Previously, the ProgressTracker only ever appended new ProgressItems to
the progress file. This made it impossible to properly save a restored
ProgressItem (i.e. a loaded ProgressItem that was ended less than the
idle period ago).

The ProgressTracker has now been revamped so that it always rewrites the
entire progress file. It does so by checking if any of the CSV lines
match any of the progress items marked as to-be-saved and replaces them
if possible.

The rewrite occurs in a new temporary file, so the original file is left
untouched until the end, where the original file is renamed into a .bak
file and the new file renamed into the progress file's filename. If any
of those file operations fail, the original file is restored.

The ProgressTracker::renameActiveFile() method has also been adjusted
and now (theoretically) rewrites the progress file in the same way, only
this time it doesn't touch anything but the file path. This has not yet
been tested though.