Remove TextEdit, inherit QQuickPaintedItem
This approach does solve the major performance issues when loading large
files (though there is still a small wait until the file is parsed).

This commit includes only rudimentary text painting into the scene. The
text is not smoothed, aliased, or otherwise optimized for display, nor
can it currently be edited.