Implement RecyclerView-like text area prototype
The prototype determines if the text block that is to be drawn is
currently visible (i.e. on screen). If it is not, it will skip it.

This results in a drastic performance increase. However, implementing
all the other functionality (like hit testing, a visible text cursor
that blinks, etc.) will not be trivial. Therefore, switching to
Qt Widgets may still be preferable.