Remove Flickable
Removing the Flickable surrounding the FormattableTextArea proved
necessary because the resulting layout was flawed. It wasn't really the
text area that was being scrolled, it was the Flickable. The text area
always stayed at 0,0 with window width and height. As a result, mouse
events didn't register outside of the text area if you scrolled down,
and if you enabled clipping, the text would only be shown fully if
unscrolled.

Since the text area definitely needs information about its scroll
position to determine which text blocks it should render, another
solution had to be found. Now the text area implements basic scrolling
behaviour and the Flickable has been removed.