Fix word count race condition
There was a race condition where, in a sufficiently large document, the
FormattableTextArea::load() function would finish before the word count
was done updating. As a result, the word count would finish updating
with m_loading set to false and the ProgressTracker would assume the new
words were legitimately added words.

The new load() function checks if the word count is still being updated
and, if so, does not reset loading to false. The word counter will then
update the loading property accordingly. This should hopefully fix the
problem.