Fix word count race condition (for real now)
There was an additional complexity I was previously unaware of. When you
emit a signal, the corresponding slots don't actually receive the signal
until the UI thread has finished all its other operations.

In concrete terms: when emitting wordCountChanged(), the corresponding
listener in QML was not actually called until both the updateWordCount()
and load() functions had finished executing, at which point loading
would be set back to false and the word count would be considered part
of the day's progress.

I've added an argument to the wordCountChanged signal that should fix
this problem.