Fix inconsistent file paths
It seems that Qt.resolvedUrl() does not natively understand file paths.
Previously that was fine because the QML url function toString() always
returned the file url with two forward slashes after the drive name
rather than one, which was then saved into the DocumentSettings.

Unfortunately this caused several issues: the file dialogs always
returned a "correct" file path with only one forward slash, which was
then stored in the application that way. If, however, you simply started
the program and continued in whatever file you were working on before,
the file url would not be retrieved from a file dialog but from the
DocumentSettings (which had two forward slashes).

In the end, this caused a problem when saving (you'd have to save twice
in order to *actually* save the file as it self-corrected itself after
the first go), but more importantly when committing progress items as
the progress items simply use whatever file path they're handed. This
meant that some paths were stored with one forward slash, some with two,
and this made matching the two impossible. Progress items that actually
belonged to the same file but had a different number of forward slashes
would not be recognized as belonging to the same file.

This has now been solved via a new global QML context object that
supplies methods for properly converting a string into a local file url
and vice versa.