Extract actions
This does not merely serve the purpose of reducing code bloat in
MainWindow.qml. There was a bug where, if the user clicked on the menu
bar just once, most actions would then require two shortcut inputs to
activate the action. This lasted until the application was restarted.
Quite bizarre. I wager that it's some kind of focus bug, but I was
unable to track it down. The official documentation shows examples where
Action elements are simply added to Menu elements like that, but it
seems to be somewhat buggy. Attaching the actions to the
ApplicationWindow instead works every time, regardless of focus.