Fix Highlighter not unhighlighting comments
Problem: When a comment bracket is placed, it immediately highlights the
rest of the current block and all subsequent blocks as comments. If a
closing bracket is then placed, all text after the comment is
unhighlighted. However, if instead of placing the closing bracket, the
opening bracket is removed, the current block is unhighlighted, but all
subsequent blocks remained highlighted. This was because the flag
EndsWithUnclosedComment was never cleared if there was no startIndex.