Implement proper scrolling-on-type behaviour
"keepScrollWhenTyping" is a user-configurable setting. The setting can be in two states.

false (default): When the contentHeight of the text area changes (i.e.
the user adds a line or paragraph or removes one), the scroll position
does not follow the cursor. That is, the cursor will continuously move
downwards if new lines are added until it reaches the bounds of the text
area.
true: When the contentHeight of the text area changes, the scroll
position follows the cursor. That is, the cursor will always be in the
same vertical position on the screen.