Simplify MainWindow background
Rather than using a Rectangle with the windowBackground color and adding
it to the scene graph, we are now just setting the
ApplicationWindow.color. This fixes an issue where, after removing the
background image from the theme, the MainWindow's background color would
revert to the default window color until restart.