Isolate functionality, loosen segment coupling
One notable problem that the FormattableTextArea had before was that it
was tightly coupled with the DocumentSegment. It still is, thanks to
methods like next() and previous(), but the updateWords() function is
now only loosely coupled with the FormattableTextArea.

Before, the DocumentSegment would call
FormattableTextArea::updateWordCount() whenever its word count changed.
Now it instead emits a wordChanged() signal (as before) and the
FormattableTextArea simply connects itself to that signal whenever a new
DocumentSegment is created.

Another major change is the removal of the signals
FormattableTextArea::textInserted() and
FormattableTextArea::textRemoved(). Instead, the function textChanged()
has now been connected to QTextDocument::contentsChange() with its three
parameters (position, removed, added). Since this signal behaves a
little differently compared to FormattableTextArea::textInserted() and
FormattableTextArea::textRemoved() (no "text" argument, just string
lengths), the DocumentSegment updating had to be adjusted slightly.