Fix scene breaks being discarded on load (#31)
The root of the problem was that insertSceneBreak() did not correctly
set the userState() property for that QTextBlock because it did not
check if userState() was previously -1 (in which case the bitwise
or won't work as expected).

This caused insertSceneBreak() to be called on a text block the function
recognized as a scene break. Since the "replaceBlock" parameter was not
passed, the default behaviour in that case is to remove the scene break.

Both of these problems have now been fixed. The function
insertSceneBreak() should now be called with "replaceBlock" passed in
the appropriate places, and the userState() property is correctly set.