Set document properties before load (#34)
Previously we set various document properties like font family, text
width etc. after the text was parsed into the document. It's unclear why
cengels' mind initially decided this was the correct course of action,
but it is immediately obvious why this is not a good idea. Setting style
properties causes a complete re-layouting pass and redraw, which is fine
if you've got an empty document, but not so fine if your document is
full of text.

Now the document is first configured and then text inserted during the
parse, making the load process over 1200% faster depending on how much
text is loaded in. Basically, the loading is no longer O(n) but closer
to O(1).