Add write lock to DocumentSegment (#36)
There was a bug where ctrl + backspacing too quickly in too short a
period of time would cause a segmentation fault because the run() block
of a DocumentSegment was called twice in parallel, causing simultaneous
access to the non-thread-safe m_words property.

The write lock automatically locks the DocumentSegment when run() is
called and unlocks it when run() has finished. If a DocumentSegment
attempts to run() itself while it is locked, that operation is added to
the QThreadPool queue instead and only executed once the lock is
released.