Set default values if theme json is missing key
There was a problem where, for instance, removing the firstLineIndent
key from a theme's JSON would, on next start, set it to 32 (which would
then be written to the theme on next save). This problem occurred
because the theme's read() method only checked if a key exists before it
gets its property value. It did not, however, set a default value in
case the key is missing.

That is a problem, because it meant that the
default values from the constructor (which are intended for the default
themes and provide incorrect "missing property" behaviour) were used
instead.