Refresh document segments on applyHeading()
Without this call, the document structure is not automatically updated
whenever the user creates a new segment (i.e. part, chapter, ...), so
the new segment would never show up in the document structure drawer.

This is hardly an optimal solution--for instance, the segment subtitle is still not
updated when it is changed until restart--but it is at least workable
for now.