Fix incorrect hex stringification
Previously, toHexadecimal() simply iterated through each QChar in a
string and converted it to hex. This is not actually a valid approach
due to the fact that a QChar can actually be multiple bytes long, for
example the double curly quote character (“), which corresponds to e2 80
9c. These kinds of characters were incorrectly converted. By using a
QByteArray, this problem no longer occurs.