Fix textArea.loading signal never called (#45)
In QML, the textArea's onLoadingChanged slot was never called, so the
scroll bar never got the notification that the text area had initialized
and it should start scrolling now.

The most likely cause is that the text area finished loading before the
QML tree was finished initializing. As a result, since the QML loading
property was internally initialized with "false", it could never become
"true". The loading property now initializes itself to textArea.loading
instead to solve this problem.