Add new paint node-based search highlighter (#32)
The new highlighter no longer inherits from QSyntaxHighlighter and
therefore no longer changes the QTextDocument's internal formats when
search matches are set or reset. Instead it is embedded directly in the
painting process of FormattableTextArea::updatePaintNode(), which means
that corresponding rectangle highlighting nodes are only added if they
are actually on-screen, which results in an enormous performance boost,
especially for search inputs with a lot of matches.