Use timestamp to subtract dates (fixes #48)
Previously the relative(date) function used the millisecond timestamp
only to calculate differences for seconds, minutes, and hours. It used a
different approach to calculate the difference in days, making use of
Date.getDate() and similar functions to determine whether a Date falls
on another day (rather than just being a full day apart).

This approach was overly complicated and error-prone. Now the function
simply uses Date.setHours() to create a Date object pointing to the
start of the current day. It then retrieves the POSIX timestamp from it
and uses this to calculate the difference in days, weeks, months, and
years.