Rework scene break into horizontal rule (#37)
Previously users could define a custom "scene break string" (like "###")
which would be used to identify scene breaks in the document and format
them accordingly.

This has now been removed. The biggest reason for the change is that it
required the program to actively check on each text change whether there
are any unformatted occurrences of said scene break string in the
document. Not only did this cause some overhead, due to a bug in the way
QTextCursor::joinPreviousEditBlock() and QTextCursor::setBlockFormat()
interact, it was not possible to merge the format change into the same
undo operation as the text change (see #37 for more info).

In this new version, scene breaks can no longer be added by simply
typing a string into the document. Instead, the "insert scene break"
command must be used, either from the toolbar menu or the shortcut. This
will insert a new QTextBlock with a scene break format applied. The
updatePaintNode() method will look for QTextBlocks with this format and
insert a horizontal rule into the node graph.

The alternative would have been to insert the horizontal rule directly
into the QTextDocument using QTextCursor::insertHtml("<hr>"), but I was
unsuccessful in getting that approach to work, possibly because of the
custom rendering process.