Always re-count the words on text change
Previously the text document would only re-count the words for a given
segment if it found a word separator in front or behind it. This
approach was flawed, though. Due to the fact that there is no way to
retrieve the character sequence that was added or removed, it is
impossible to accurately determine whether the text change included a
word separator or not.

By always re-counting the words, for instance, opening and closing
comments now always count correctly.

This is likely to have some performance impact, but initial tests did
not see an immediate problem. Any further problems will likely be
resolved by #38.