Change Home/End behaviour
The Home and End keys now conform to common text processing behaviour.
They move to the start/end of the block if CTRL is not held and to the
start/end of the document if CTRL is held. They also now move the caret.

It is debatable whether the Page Up/Down keys should also move the
caret. This might be annoying though if one just seeks to quickly look
at the last page of text without moving the caret.