Make Sky.Text PlainText by default
After some tests, it turns out that there is a significant performance
impact on all text elements that combine wrapMode: Text.Wrap with
textFormat: Text.RichText. A window with many such text elements will
take over a second to initialize and show its contents, even if none of
the text contains any formatting.

As a result, Sky.Text now defaults to plain text and can conditionally
be turned to rich text if needed.