Move word count information into UserData
Previously word counting worked like this:
user enters/removes text -> find the DocumentSegment(s) in which the
change occurred -> DocumentSegment(s) recount all words contained within

This resulted in an unnecessarily heavy approach. The burden of word
counting has been moved from the DocumentSegment to the QTextBlock, so
now it works like this:
user enters/removes text -> find the QTextBlock(s) in which the change
occurred -> QTextBlock(s) recount all words contained within ->
DocumentSegment(s) in which the change occurred iterate through all
contained QTextBlocks and accumulate the new word count

This represents an enormous performance improvement as typing a single
character or removing it (as is most often the case) no longer results
in the entire DocumentSegment recounting its words but only a single
QTextBlock.

This solution has not been widely tested yet. There still be some
inconsistencies where the DocumentSegments and the QTextBlocks go out of
sync or similar issues.