Improve menuBarItem responsiveness
The default behaviour of menuBarItems, when clicked, apparently follows
the menu.opened property. This property is only set to true after a
slight delay, regardless of any transitions used. This means that, by
default, clicking on a menuBarItem twice in rapid succession will just
ignore all subsequent clicks.

The new approach handles this more appropriately and makes sure that all
clicks are respected.