Improve word count algorithm, remove worker thread
Previously the word counting algorithm used asynchronicity to ensure the
word counting process would not block the UI thread (and thereby user
input). However, this of course also resulted in staggered updates to
the visible word counts, making the overall user experience worse.

In this new approach the word counting algorithm has been improved to
such a degree as to be fast enough not to block user input even on
extremely large files. As a result, concurrency is no longer required
for the word counting and has been removed.

As a nice side-effect, the TextIterator now relies on UserData
information computed by the TextFormatter to determine whether a word is
part of a comment instead of trying to pattern match comments itself.
This resolves some issues where the TextIterator's text range lies
inside of a comment range, thus having insufficient information to
properly detect the comment.