Improve updateDocumentDefaults() performance
The function no longer conditionally operates on a cloned version of the
QTextDocument to avoid signals being emitted mid-update. The performance
of that was lacking.

I first experimented with simply disconnecting its signals and
reconnecting them at the end, but that still wasn't good enough. The
problem isn't the signals; it's the fact that the entire document layout
has to be recomputed when changing these values. This is a bottleneck in
Qt's Scribe framework and I can only work around it, not fix it.

The workaround, in this case, is to simply empty the QTextDocument of
all text and then parse the text back in after the styling has been
updated. This makes the style updates instant, but the parsing still
takes a not-insignificant amount of time. It's already a huge
improvement though when switching themes.

It's still not an ideal solution though because doing this loses the
document structure and word counts, both of which have to be
recalculated afterwards, resulting in an additional performance hit.
Ideally there would be a way to copy one QTextDocument into another
while keeping its blocks and user data intact--which would eliminate
both the reparsing and word counting--but that doesn't appear to be
possible.