Simplify and optimize updateDocumentStructure()
Previously the function was 70 lines long. By utilizing pre-existing
functions and simplifying the logic, this has been reduced to about 30.

This also fixed a bug where the document structure would enter an
inconsistent state when updateDocumentStructure() was called with
unreasonable arguments (which QTextDocument sometimes outputs for
unknown reasons).

Other changes:
* use std::find_if() in findDocumentSegment()